Refactored some functions from player_history.py to blacklist.py and queries.py
