+++ rconweb/api/views.py
@@ -17,11 +17,13 @@ from rcon.commands import CommandFailedError
 from rcon.settings import SERVER_INFO
 from rcon.player_history import (
     get_players_by_appearance, 
-    add_player_to_blacklist, 
-    remove_player_from_blacklist,
     get_player_profile,
     add_flag_to_player,
-    remove_flag,
+    remove_flag
+)
+from rcon.blacklist import (
+    add_player_to_blacklist, 
+    remove_player_from_blacklist,
     banned_unblacklisted_players,
     unban_unblacklisted_players as pv_unban_unblacklisted_players
 )
@@ -226,7 +228,7 @@ def blacklist_player(request):
         )
         send_to_discord_audit("Blacklist '{}' for '{}'".format(data['steam_id_64'], data['reason']), get_client_ip(request))
         name = data['name'] if 'name' in data else None
-        add_player_to_blacklist(data['steam_id_64'], data['reason'], name, ctl, True)
+        add_player_to_blacklist(data['steam_id_64'], data['reason'], name)
         failed = False
     except:
         logger.exception("Unable to blacklist player")
